+++ bson/__init__.py
@@ -76,6 +76,8 @@ from typing import (
     List,
     Mapping,
     MutableMapping,
+    NoReturn,
+    Optional,
     Sequence,
     Tuple,
     Type,
@@ -95,7 +97,12 @@ from bson.binary import (  # noqa: F401
     UuidRepresentation,
 )
 from bson.code import Code
-from bson.codec_options import DEFAULT_CODEC_OPTIONS, CodecOptions, _raw_document_class
+from bson.codec_options import (
+    DEFAULT_CODEC_OPTIONS,
+    CodecOptions,
+    _DocumentType,
+    _raw_document_class,
+)
 from bson.dbref import DBRef
 from bson.decimal128 import Decimal128
 from bson.errors import InvalidBSON, InvalidDocument, InvalidStringData
@@ -108,12 +115,10 @@ from bson.son import RE_TYPE, SON
 from bson.timestamp import Timestamp
 from bson.tz_util import utc
 
-# Import RawBSONDocument for type-checking only to avoid circular dependency.
+# Import some modules for type-checking only.
 if TYPE_CHECKING:
     from array import array
     from mmap import mmap
-
-    from bson.raw_bson import RawBSONDocument
 
 
 try:
@@ -166,7 +171,7 @@ def get_data_and_view(data: Any) -> Tuple[Any, memoryview]:
     return view.tobytes(), view
 
 
-def _raise_unknown_type(element_type: int, element_name: str) -> None:
+def _raise_unknown_type(element_type: int, element_name: str) -> NoReturn:
     """Unknown type helper."""
     raise InvalidBSON(
         "Detected unknown BSON type %r for fieldname '%s'. Are "
@@ -181,7 +186,7 @@ def _get_int(
     return _UNPACK_INT_FROM(data, position)[0], position + 4
 
 
-def _get_c_string(data: Any, view: Any, position: int, opts: Any) -> Tuple[str, int]:
+def _get_c_string(data: Any, view: Any, position: int, opts: CodecOptions) -> Tuple[str, int]:
     """Decode a BSON 'C' string to python str."""
     end = data.index(b"\x00", position)
     return _utf_8_decode(view[position:end], opts.unicode_decode_error_handler, True)[0], end + 1
@@ -195,7 +200,7 @@ def _get_float(
 
 
 def _get_string(
-    data: Any, view: Any, position: int, obj_end: int, opts: Any, dummy: Any
+    data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions, dummy: Any
 ) -> Tuple[str, int]:
     """Decode a BSON string to python str."""
     length = _UNPACK_INT_FROM(data, position)[0]
@@ -226,7 +231,7 @@ def _get_object_size(data: Any, position: int, obj_end: int) -> Tuple[int, int]:
 
 
 def _get_object(
-    data: Any, view: Any, position: int, obj_end: int, opts: Any, dummy: Any
+    data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions, dummy: Any
 ) -> Tuple[Any, int]:
     """Decode a BSON subdocument to opts.document_class or bson.dbref.DBRef."""
     obj_size, end = _get_object_size(data, position, obj_end)
@@ -247,7 +252,7 @@ def _get_object(
 
 
 def _get_array(
-    data: Any, view: Any, position: int, obj_end: int, opts: Any, element_name: str
+    data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions, element_name: str
 ) -> Tuple[Any, int]:
     """Decode a BSON array to python list."""
     size = _UNPACK_INT_FROM(data, position)[0]
@@ -289,7 +294,7 @@ def _get_array(
 
 
 def _get_binary(
-    data: Any, view: Any, position: int, obj_end: int, opts: Any, dummy1: Any
+    data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions, dummy1: Any
 ) -> Tuple[Union[Binary, uuid.UUID], int]:
     """Decode a BSON binary to bson.binary.Binary or python UUID."""
     length, subtype = _UNPACK_LENGTH_SUBTYPE_FROM(data, position)
@@ -347,14 +352,14 @@ def _get_boolean(
 
 
 def _get_date(
-    data: Any, view: Any, position: int, dummy0: int, opts: Any, dummy1: Any
+    data: Any, view: Any, position: int, dummy0: int, opts: CodecOptions, dummy1: Any
 ) -> Tuple[datetime.datetime, int]:
     """Decode a BSON datetime to python datetime.datetime."""
     return _millis_to_datetime(_UNPACK_LONG_FROM(data, position)[0], opts), position + 8
 
 
 def _get_code(
-    data: Any, view: Any, position: int, obj_end: int, opts: Any, element_name: str
+    data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions, element_name: str
 ) -> Tuple[Code, int]:
     """Decode a BSON code to bson.code.Code."""
     code, position = _get_string(data, view, position, obj_end, opts, element_name)
@@ -362,7 +367,7 @@ def _get_code(
 
 
 def _get_code_w_scope(
-    data: Any, view: Any, position: int, obj_end: int, opts: Any, element_name: str
+    data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions, element_name: str
 ) -> Tuple[Code, int]:
     """Decode a BSON code_w_scope to bson.code.Code."""
     code_end = position + _UNPACK_INT_FROM(data, position)[0]
@@ -374,7 +379,7 @@ def _get_code_w_scope(
 
 
 def _get_regex(
-    data: Any, view: Any, position: int, dummy0: Any, opts: Any, dummy1: Any
+    data: Any, view: Any, position: int, dummy0: Any, opts: CodecOptions, dummy1: Any
 ) -> Tuple[Regex, int]:
     """Decode a BSON regex to bson.regex.Regex or a python pattern object."""
     pattern, position = _get_c_string(data, view, position, opts)
@@ -384,7 +389,7 @@ def _get_regex(
 
 
 def _get_ref(
-    data: Any, view: Any, position: int, obj_end: int, opts: Any, element_name: str
+    data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions, element_name: str
 ) -> Tuple[DBRef, int]:
     """Decode (deprecated) BSON DBPointer to bson.dbref.DBRef."""
     collection, position = _get_string(data, view, position, obj_end, opts, element_name)
@@ -448,12 +453,16 @@ _ELEMENT_GETTER: Dict[int, Callable[..., Tuple[Any, int]]] = {
 
 if _USE_C:
 
-    def _element_to_dict(data: Any, view: Any, position: int, obj_end: int, opts: Any) -> Any:
+    def _element_to_dict(
+        data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions
+    ) -> Any:
         return _cbson._element_to_dict(data, position, obj_end, opts)
 
 else:
 
-    def _element_to_dict(data: Any, view: Any, position: int, obj_end: int, opts: Any) -> Any:
+    def _element_to_dict(
+        data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions
+    ) -> Any:
         """Decode a single key, value pair."""
         element_type = data[position]
         position += 1
@@ -476,13 +485,13 @@ else:
 _T = TypeVar("_T", bound=MutableMapping[Any, Any])
 
 
-def _raw_to_dict(data: Any, position: int, obj_end: int, opts: Any, result: _T) -> _T:
+def _raw_to_dict(data: Any, position: int, obj_end: int, opts: CodecOptions, result: _T) -> _T:
     data, view = get_data_and_view(data)
     return _elements_to_dict(data, view, position, obj_end, opts, result)
 
 
 def _elements_to_dict(
-    data: Any, view: Any, position: int, obj_end: int, opts: Any, result: Any = None
+    data: Any, view: Any, position: int, obj_end: int, opts: CodecOptions, result: Any = None
 ) -> Any:
     """Decode a BSON document into result."""
     if result is None:
@@ -496,7 +505,7 @@ def _elements_to_dict(
     return result
 
 
-def _bson_to_dict(data: Any, opts: Any) -> Any:
+def _bson_to_dict(data: Any, opts: CodecOptions) -> Any:
     """Decode a BSON string to document_class."""
     data, view = get_data_and_view(data)
     try:
@@ -586,7 +595,7 @@ def _encode_bytes(name: bytes, value: bytes, dummy0: Any, dummy1: Any) -> bytes:
     return b"\x05" + name + _PACK_INT(len(value)) + b"\x00" + value
 
 
-def _encode_mapping(name: bytes, value: Any, check_keys: bool, opts: Any) -> bytes:
+def _encode_mapping(name: bytes, value: Any, check_keys: bool, opts: CodecOptions) -> bytes:
     """Encode a mapping type."""
     if _raw_document_class(value):
         return b"\x03" + name + value.raw
@@ -594,7 +603,7 @@ def _encode_mapping(name: bytes, value: Any, check_keys: bool, opts: Any) -> byt
     return b"\x03" + name + _PACK_INT(len(data) + 5) + data + b"\x00"
 
 
-def _encode_dbref(name: bytes, value: DBRef, check_keys: bool, opts: Any) -> bytes:
+def _encode_dbref(name: bytes, value: DBRef, check_keys: bool, opts: CodecOptions) -> bytes:
     """Encode bson.dbref.DBRef."""
     buf = bytearray(b"\x03" + name + b"\x00\x00\x00\x00")
     begin = len(buf) - 4
@@ -611,7 +620,7 @@ def _encode_dbref(name: bytes, value: DBRef, check_keys: bool, opts: Any) -> byt
     return bytes(buf)
 
 
-def _encode_list(name: bytes, value: Sequence[Any], check_keys: bool, opts: Any) -> bytes:
+def _encode_list(name: bytes, value: Sequence[Any], check_keys: bool, opts: CodecOptions) -> bytes:
     """Encode a list/tuple."""
     lname = gen_list_name()
     data = b"".join([_name_value_to_bson(next(lname), item, check_keys, opts) for item in value])
@@ -620,8 +629,8 @@ def _encode_list(name: bytes, value: Sequence[Any], check_keys: bool, opts: Any)
 
 def _encode_text(name: bytes, value: str, dummy0: Any, dummy1: Any) -> bytes:
     """Encode a python str."""
-    value = _utf_8_encode(value)[0]
-    return b"\x02" + name + _PACK_INT(len(value) + 1) + value + b"\x00"  # type: ignore
+    bvalue = _utf_8_encode(value)[0]
+    return b"\x02" + name + _PACK_INT(len(bvalue) + 1) + bvalue + b"\x00"
 
 
 def _encode_binary(name: bytes, value: Binary, dummy0: Any, dummy1: Any) -> bytes:
@@ -632,7 +641,7 @@ def _encode_binary(name: bytes, value: Binary, dummy0: Any, dummy1: Any) -> byte
     return b"\x05" + name + _PACK_LENGTH_SUBTYPE(len(value), subtype) + value
 
 
-def _encode_uuid(name: bytes, value: uuid.UUID, dummy: Any, opts: Any) -> bytes:
+def _encode_uuid(name: bytes, value: uuid.UUID, dummy: Any, opts: CodecOptions) -> bytes:
     """Encode uuid.UUID."""
     uuid_representation = opts.uuid_representation
     binval = Binary.from_uuid(value, uuid_representation=uuid_representation)
@@ -686,7 +695,7 @@ def _encode_regex(name: bytes, value: Regex, dummy0: Any, dummy1: Any) -> bytes:
         return b"\x0B" + name + _make_c_string_check(value.pattern) + sflags
 
 
-def _encode_code(name: bytes, value: Code, dummy: Any, opts: Any) -> bytes:
+def _encode_code(name: bytes, value: Code, dummy: Any, opts: CodecOptions) -> bytes:
     """Encode bson.code.Code."""
     cstring = _make_c_string(value)
     cstrlen = len(cstring)
@@ -790,7 +799,7 @@ def _name_value_to_bson(
     name: bytes,
     value: Any,
     check_keys: bool,
-    opts: Any,
+    opts: CodecOptions,
     in_custom_call: bool = False,
     in_fallback_call: bool = False,
 ) -> bytes:
@@ -843,7 +852,7 @@ def _name_value_to_bson(
     raise InvalidDocument("cannot encode object: %r, of type: %r" % (value, type(value)))
 
 
-def _element_to_bson(key: Any, value: Any, check_keys: bool, opts: Any) -> bytes:
+def _element_to_bson(key: Any, value: Any, check_keys: bool, opts: CodecOptions) -> bytes:
     """Encode a single key, value pair."""
     if not isinstance(key, str):
         raise InvalidDocument("documents must have only string keys, key was %r" % (key,))
@@ -857,7 +866,7 @@ def _element_to_bson(key: Any, value: Any, check_keys: bool, opts: Any) -> bytes
     return _name_value_to_bson(name, value, check_keys, opts)
 
 
-def _dict_to_bson(doc: Any, check_keys: bool, opts: Any, top_level: bool = True) -> bytes:
+def _dict_to_bson(doc: Any, check_keys: bool, opts: CodecOptions, top_level: bool = True) -> bytes:
     """Encode a document to BSON."""
     if _raw_document_class(doc):
         return cast(bytes, doc.raw)
@@ -879,7 +888,7 @@ if _USE_C:
     _dict_to_bson = _cbson._dict_to_bson  # noqa: F811
 
 
-def _millis_to_datetime(millis: int, opts: Any) -> datetime.datetime:
+def _millis_to_datetime(millis: int, opts: CodecOptions) -> datetime.datetime:
     """Convert milliseconds since epoch UTC to datetime."""
     diff = ((millis % 1000) + 1000) % 1000
     seconds = (millis - diff) // 1000
@@ -904,7 +913,6 @@ _CODEC_OPTIONS_TYPE_ERROR = TypeError("codec_options must be an instance of Code
 
 
 _DocumentIn = Mapping[str, Any]
-_DocumentOut = Union[MutableMapping[str, Any], "RawBSONDocument"]
 _ReadableBuffer = Union[bytes, memoryview, "mmap", "array"]
 
 
@@ -940,8 +948,8 @@ def encode(
 
 
 def decode(
-    data: _ReadableBuffer, codec_options: CodecOptions = DEFAULT_CODEC_OPTIONS
-) -> Dict[str, Any]:
+    data: _ReadableBuffer, codec_options: "Optional[CodecOptions[_DocumentType]]" = None
+) -> _DocumentType:
     """Decode BSON to a document.
 
     By default, returns a BSON document represented as a Python
@@ -967,15 +975,16 @@ def decode(
 
     .. versionadded:: 3.9
     """
-    if not isinstance(codec_options, CodecOptions):
+    opts: CodecOptions = codec_options or DEFAULT_CODEC_OPTIONS
+    if not isinstance(opts, CodecOptions):
         raise _CODEC_OPTIONS_TYPE_ERROR
 
-    return _bson_to_dict(data, codec_options)
+    return _bson_to_dict(data, opts)
 
 
 def decode_all(
-    data: _ReadableBuffer, codec_options: CodecOptions = DEFAULT_CODEC_OPTIONS
-) -> List[Dict[str, Any]]:
+    data: _ReadableBuffer, codec_options: "Optional[CodecOptions[_DocumentType]]" = None
+) -> List[_DocumentType]:
     """Decode BSON data to multiple documents.
 
     `data` must be a bytes-like object implementing the buffer protocol that
@@ -998,15 +1007,16 @@ def decode_all(
        Replaced `as_class`, `tz_aware`, and `uuid_subtype` options with
        `codec_options`.
     """
+    opts = codec_options or DEFAULT_CODEC_OPTIONS
     data, view = get_data_and_view(data)
-    if not isinstance(codec_options, CodecOptions):
+    if not isinstance(opts, CodecOptions):
         raise _CODEC_OPTIONS_TYPE_ERROR
 
     data_len = len(data)
-    docs = []
+    docs: List[_DocumentType] = []
     position = 0
     end = data_len - 1
-    use_raw = _raw_document_class(codec_options.document_class)
+    use_raw = _raw_document_class(opts.document_class)
     try:
         while position < end:
             obj_size = _UNPACK_INT_FROM(data, position)[0]
@@ -1017,10 +1027,10 @@ def decode_all(
                 raise InvalidBSON("bad eoo")
             if use_raw:
                 docs.append(
-                    codec_options.document_class(data[position : obj_end + 1], codec_options)
+                    opts.document_class(data[position : obj_end + 1], codec_options)  # type: ignore
                 )
             else:
-                docs.append(_elements_to_dict(data, view, position + 4, obj_end, codec_options))
+                docs.append(_elements_to_dict(data, view, position + 4, obj_end, opts))
             position += obj_size
         return docs
     except InvalidBSON:
@@ -1110,8 +1120,8 @@ def _decode_all_selective(data: Any, codec_options: CodecOptions, fields: Any) -
 
 
 def decode_iter(
-    data: bytes, codec_options: CodecOptions = DEFAULT_CODEC_OPTIONS
-) -> Iterator[_DocumentOut]:
+    data: bytes, codec_options: "Optional[CodecOptions[_DocumentType]]" = None
+) -> Iterator[_DocumentType]:
     """Decode BSON data to multiple documents as a generator.
 
     Works similarly to the decode_all function, but yields one document at a
@@ -1131,7 +1141,8 @@ def decode_iter(
 
     .. versionadded:: 2.8
     """
-    if not isinstance(codec_options, CodecOptions):
+    opts = codec_options or DEFAULT_CODEC_OPTIONS
+    if not isinstance(opts, CodecOptions):
         raise _CODEC_OPTIONS_TYPE_ERROR
 
     position = 0
@@ -1141,12 +1152,12 @@ def decode_iter(
         elements = data[position : position + obj_size]
         position += obj_size
 
-        yield _bson_to_dict(elements, codec_options)
+        yield _bson_to_dict(elements, opts)
 
 
 def decode_file_iter(
-    file_obj: Union[BinaryIO, IO], codec_options: CodecOptions = DEFAULT_CODEC_OPTIONS
-) -> Iterator[_DocumentOut]:
+    file_obj: Union[BinaryIO, IO], codec_options: "Optional[CodecOptions[_DocumentType]]" = None
+) -> Iterator[_DocumentType]:
     """Decode bson data from a file to multiple documents as a generator.
 
     Works similarly to the decode_all function, but reads from the file object
@@ -1163,6 +1174,7 @@ def decode_file_iter(
 
     .. versionadded:: 2.8
     """
+    opts = codec_options or DEFAULT_CODEC_OPTIONS
     while True:
         # Read size of next object.
         size_data = file_obj.read(4)
@@ -1172,7 +1184,7 @@ def decode_file_iter(
             raise InvalidBSON("cut off in middle of objsize")
         obj_size = _UNPACK_INT_FROM(size_data, 0)[0] - 4
         elements = size_data + file_obj.read(max(0, obj_size))
-        yield _bson_to_dict(elements, codec_options)
+        yield _bson_to_dict(elements, opts)
 
 
 def is_valid(bson: bytes) -> bool:
@@ -1233,7 +1245,7 @@ class BSON(bytes):
         """
         return cls(encode(document, check_keys, codec_options))
 
-    def decode(self, codec_options: CodecOptions = DEFAULT_CODEC_OPTIONS) -> Dict[str, Any]:  # type: ignore[override]
+    def decode(self, codec_options: CodecOptions = DEFAULT_CODEC_OPTIONS) -> _DocumentType:  # type: ignore[override]
         """Decode this BSON data.
 
         By default, returns a BSON document represented as a Python