+++ pymongo/collection.py
@@ -23,6 +23,7 @@ from typing import (
     List,
     Mapping,
     MutableMapping,
+    NoReturn,
     Optional,
     Sequence,
     Tuple,
@@ -343,7 +344,7 @@ class Collection(common.BaseObject, Generic[_DocumentType]):
     def __hash__(self) -> int:
         return hash((self.__database, self.__name))
 
-    def __bool__(self) -> bool:
+    def __bool__(self) -> NoReturn:
         raise NotImplementedError(
             "Collection objects do not implement truth "
             "value testing or bool(). Please compare "
@@ -1882,8 +1883,8 @@ class Collection(common.BaseObject, Generic[_DocumentType]):
 
         .. _$expr: https://docs.mongodb.com/manual/reference/operator/query/expr/
         .. _$geoWithin: https://docs.mongodb.com/manual/reference/operator/query/geoWithin/
-        .. _$center: https://docs.mongodb.com/manual/reference/operator/query/center/#op._S_center
-        .. _$centerSphere: https://docs.mongodb.com/manual/reference/operator/query/centerSphere/#op._S_centerSphere
+        .. _$center: https://docs.mongodb.com/manual/reference/operator/query/center/
+        .. _$centerSphere: https://docs.mongodb.com/manual/reference/operator/query/centerSphere/
         """
         pipeline = [{"$match": filter}]
         if "skip" in kwargs:
@@ -2109,7 +2110,7 @@ class Collection(common.BaseObject, Generic[_DocumentType]):
 
         .. seealso:: The MongoDB documentation on `indexes <https://dochub.mongodb.org/core/indexes>`_.
 
-        .. _wildcard index: https://docs.mongodb.com/master/core/index-wildcard/#wildcard-index-core
+        .. _wildcard index: https://docs.mongodb.com/master/core/index-wildcard/
         """
         cmd_options = {}
         if "maxTimeMS" in kwargs:
@@ -2254,7 +2255,7 @@ class Collection(common.BaseObject, Generic[_DocumentType]):
 
         .. versionadded:: 3.0
         """
-        codec_options = CodecOptions(SON)
+        codec_options: CodecOptions = CodecOptions(SON)
         coll = self.with_options(
             codec_options=codec_options, read_preference=ReadPreference.PRIMARY
         )
@@ -2655,7 +2656,7 @@ class Collection(common.BaseObject, Generic[_DocumentType]):
 
         .. versionadded:: 3.6
 
-        .. seealso:: The MongoDB documentation on `changeStreams <https://dochub.mongodb.org/core/changeStreams>`_.
+        .. seealso:: The MongoDB documentation on `changeStreams <https://docs.mongodb.com/manual/changeStreams/>`_.
 
         .. _change streams specification:
             https://github.com/mongodb/specifications/blob/master/source/change-streams/change-streams.rst
@@ -3241,12 +3242,12 @@ class Collection(common.BaseObject, Generic[_DocumentType]):
     def __iter__(self) -> "Collection[_DocumentType]":
         return self
 
-    def __next__(self) -> None:
+    def __next__(self) -> NoReturn:
         raise TypeError("'Collection' object is not iterable")
 
     next = __next__
 
-    def __call__(self, *args: Any, **kwargs: Any) -> None:
+    def __call__(self, *args: Any, **kwargs: Any) -> NoReturn:
         """This is only here so that some API misusages are easier to debug."""
         if "." not in self.__name:
             raise TypeError(