+++ test/test_mypy.py
@@ -16,8 +16,9 @@
 sample client code that uses PyMongo typings."""
 
 import os
+import tempfile
 import unittest
-from typing import TYPE_CHECKING, Any, Dict, Iterable, List
+from typing import TYPE_CHECKING, Any, Dict, Iterable, Iterator, List
 
 try:
     from typing import TypedDict  # type: ignore[attr-defined]
@@ -39,6 +40,7 @@ except ImportError:
 from test import IntegrationTest
 from test.utils import rs_or_single_client
 
+from bson import CodecOptions, decode, decode_all, decode_file_iter, decode_iter, encode
 from bson.raw_bson import RawBSONDocument
 from bson.son import SON
 from pymongo.collection import Collection
@@ -52,6 +54,15 @@ def get_tests() -> Iterable[str]:
     for dirpath, _, filenames in os.walk(TEST_PATH):
         for filename in filenames:
             yield os.path.join(dirpath, filename)
+
+
+def only_type_check(func):
+    def inner(*args, **kwargs):
+        if not TYPE_CHECKING:
+            raise unittest.SkipTest("Used for Type Checking Only")
+        func(*args, **kwargs)
+
+    return inner
 
 
 class TestMypyFails(unittest.TestCase):
@@ -105,7 +116,7 @@ class TestPymongo(IntegrationTest):
         self.assertTrue(result.acknowledged)
 
     def test_command(self) -> None:
-        result = self.client.admin.command("ping")
+        result: Dict = self.client.admin.command("ping")
         items = result.items()
 
     def test_list_collections(self) -> None:
@@ -124,43 +135,6 @@ class TestPymongo(IntegrationTest):
         doc = {"my": "doc"}
         coll.insert_one(doc)
         retreived = coll.find_one({"_id": doc["_id"]})
-        assert retreived is not None
-        retreived["a"] = 1
-
-    def test_explicit_document_type(self) -> None:
-        if not TYPE_CHECKING:
-            raise unittest.SkipTest("Do not use raw MongoClient")
-        client: MongoClient[Dict[str, Any]] = MongoClient()
-        coll = client.test.test
-        retreived = coll.find_one({"_id": "foo"})
-        assert retreived is not None
-        retreived["a"] = 1
-
-    def test_typeddict_document_type(self) -> None:
-        if not TYPE_CHECKING:
-            raise unittest.SkipTest("Do not use raw MongoClient")
-        client: MongoClient[Movie] = MongoClient()
-        coll = client.test.test
-        retreived = coll.find_one({"_id": "foo"})
-        assert retreived is not None
-        assert retreived["year"] == 1
-        assert retreived["name"] == "a"
-
-    def test_raw_bson_document_type(self) -> None:
-        if not TYPE_CHECKING:
-            raise unittest.SkipTest("Do not use raw MongoClient")
-        client = MongoClient(document_class=RawBSONDocument)
-        coll = client.test.test
-        retreived = coll.find_one({"_id": "foo"})
-        assert retreived is not None
-        assert len(retreived.raw) > 0
-
-    def test_son_document_type(self) -> None:
-        if not TYPE_CHECKING:
-            raise unittest.SkipTest("Do not use raw MongoClient")
-        client = MongoClient(document_class=SON[str, Any])
-        coll = client.test.test
-        retreived = coll.find_one({"_id": "foo"})
         assert retreived is not None
         retreived["a"] = 1
 
@@ -186,6 +160,221 @@ class TestPymongo(IntegrationTest):
             ]
         )
         self.assertTrue(len(list(result)))
+
+
+class TestDecode(unittest.TestCase):
+    def test_bson_decode(self) -> None:
+        doc = {"_id": 1}
+        bsonbytes = encode(doc)
+        rt_document: Dict[str, Any] = decode(bsonbytes)
+        assert rt_document["_id"] == 1
+        rt_document["foo"] = "bar"
+
+        class MyDict(Dict[str, Any]):
+            def foo(self):
+                return "bar"
+
+        codec_options = CodecOptions(document_class=MyDict)
+        bsonbytes2 = encode(doc, codec_options=codec_options)
+        rt_document2 = decode(bsonbytes2, codec_options=codec_options)
+        assert rt_document2.foo() == "bar"
+
+        codec_options2 = CodecOptions(document_class=RawBSONDocument)
+        bsonbytes3 = encode(doc, codec_options=codec_options2)
+        rt_document3 = decode(bsonbytes2, codec_options=codec_options2)
+        assert rt_document3.raw
+
+    def test_bson_decode_all(self) -> None:
+        doc = {"_id": 1}
+        bsonbytes = encode(doc)
+        bsonbytes += encode(doc)
+        rt_documents: List[Dict[str, Any]] = decode_all(bsonbytes)
+        assert rt_documents[0]["_id"] == 1
+        rt_documents[0]["foo"] = "bar"
+
+        class MyDict(Dict[str, Any]):
+            def foo(self):
+                return "bar"
+
+        codec_options2 = CodecOptions(MyDict)
+        bsonbytes2 = encode(doc, codec_options=codec_options2)
+        bsonbytes2 += encode(doc, codec_options=codec_options2)
+        rt_documents2 = decode_all(bsonbytes2, codec_options2)
+        assert rt_documents2[0].foo() == "bar"
+
+        codec_options3 = CodecOptions(RawBSONDocument)
+        bsonbytes3 = encode(doc, codec_options=codec_options3)
+        bsonbytes3 += encode(doc, codec_options=codec_options3)
+        rt_documents3 = decode_all(bsonbytes3, codec_options3)
+        assert rt_documents3[0].raw
+
+    def test_bson_decode_iter(self) -> None:
+        doc = {"_id": 1}
+        bsonbytes = encode(doc)
+        bsonbytes += encode(doc)
+        rt_documents: Iterator[Dict[str, Any]] = decode_iter(bsonbytes)
+        assert next(rt_documents)["_id"] == 1
+        next(rt_documents)["foo"] = "bar"
+
+        class MyDict(Dict[str, Any]):
+            def foo(self):
+                return "bar"
+
+        codec_options2 = CodecOptions(MyDict)
+        bsonbytes2 = encode(doc, codec_options=codec_options2)
+        bsonbytes2 += encode(doc, codec_options=codec_options2)
+        rt_documents2 = decode_iter(bsonbytes2, codec_options2)
+        assert next(rt_documents2).foo() == "bar"
+
+        codec_options3 = CodecOptions(RawBSONDocument)
+        bsonbytes3 = encode(doc, codec_options=codec_options3)
+        bsonbytes3 += encode(doc, codec_options=codec_options3)
+        rt_documents3 = decode_iter(bsonbytes3, codec_options3)
+        assert next(rt_documents3).raw
+
+    def make_tempfile(self, content: bytes) -> Any:
+        fileobj = tempfile.TemporaryFile()
+        fileobj.write(content)
+        fileobj.seek(0)
+        self.addCleanup(fileobj.close)
+        return fileobj
+
+    def test_bson_decode_file_iter(self) -> None:
+        doc = {"_id": 1}
+        bsonbytes = encode(doc)
+        bsonbytes += encode(doc)
+        fileobj = self.make_tempfile(bsonbytes)
+        rt_documents: Iterator[Dict[str, Any]] = decode_file_iter(fileobj)
+        assert next(rt_documents)["_id"] == 1
+        next(rt_documents)["foo"] = "bar"
+
+        class MyDict(Dict[str, Any]):
+            def foo(self):
+                return "bar"
+
+        codec_options2 = CodecOptions(MyDict)
+        bsonbytes2 = encode(doc, codec_options=codec_options2)
+        bsonbytes2 += encode(doc, codec_options=codec_options2)
+        fileobj2 = self.make_tempfile(bsonbytes2)
+        rt_documents2 = decode_file_iter(fileobj2, codec_options2)
+        assert next(rt_documents2).foo() == "bar"
+
+        codec_options3 = CodecOptions(RawBSONDocument)
+        bsonbytes3 = encode(doc, codec_options=codec_options3)
+        bsonbytes3 += encode(doc, codec_options=codec_options3)
+        fileobj3 = self.make_tempfile(bsonbytes3)
+        rt_documents3 = decode_file_iter(fileobj3, codec_options3)
+        assert next(rt_documents3).raw
+
+
+class TestDocumentType(unittest.TestCase):
+    @only_type_check
+    def test_default(self) -> None:
+        client: MongoClient = MongoClient()
+        coll = client.test.test
+        retreived = coll.find_one({"_id": "foo"})
+        assert retreived is not None
+        retreived["a"] = 1
+
+    @only_type_check
+    def test_explicit_document_type(self) -> None:
+        client: MongoClient[Dict[str, Any]] = MongoClient()
+        coll = client.test.test
+        retreived = coll.find_one({"_id": "foo"})
+        assert retreived is not None
+        retreived["a"] = 1
+
+    @only_type_check
+    def test_typeddict_document_type(self) -> None:
+        client: MongoClient[Movie] = MongoClient()
+        coll = client.test.test
+        retreived = coll.find_one({"_id": "foo"})
+        assert retreived is not None
+        assert retreived["year"] == 1
+        assert retreived["name"] == "a"
+
+    @only_type_check
+    def test_raw_bson_document_type(self) -> None:
+        client = MongoClient(document_class=RawBSONDocument)
+        coll = client.test.test
+        retreived = coll.find_one({"_id": "foo"})
+        assert retreived is not None
+        assert len(retreived.raw) > 0
+
+    @only_type_check
+    def test_son_document_type(self) -> None:
+        client = MongoClient(document_class=SON[str, Any])
+        coll = client.test.test
+        retreived = coll.find_one({"_id": "foo"})
+        assert retreived is not None
+        retreived["a"] = 1
+
+
+class TestCommandDocumentType(unittest.TestCase):
+    @only_type_check
+    def test_default(self) -> None:
+        client: MongoClient = MongoClient()
+        result: Dict = client.admin.command("ping")
+        result["a"] = 1
+
+    @only_type_check
+    def test_explicit_document_type(self) -> None:
+        client: MongoClient = MongoClient()
+        codec_options: CodecOptions[Dict[str, Any]] = CodecOptions()
+        result = client.admin.command("ping", codec_options=codec_options)
+        result["a"] = 1
+
+    @only_type_check
+    def test_typeddict_document_type(self) -> None:
+        client: MongoClient = MongoClient()
+        codec_options: CodecOptions[Movie] = CodecOptions()
+        result = client.admin.command("ping", codec_options=codec_options)
+        assert result["year"] == 1
+        assert result["name"] == "a"
+
+    @only_type_check
+    def test_raw_bson_document_type(self) -> None:
+        client: MongoClient = MongoClient()
+        codec_options = CodecOptions(RawBSONDocument)
+        result = client.admin.command("ping", codec_options=codec_options)
+        assert len(result.raw) > 0
+
+    @only_type_check
+    def test_son_document_type(self) -> None:
+        client = MongoClient(document_class=SON[str, Any])
+        codec_options = CodecOptions(SON[str, Any])
+        result = client.admin.command("ping", codec_options=codec_options)
+        result["a"] = 1
+
+
+class TestCodecOptionsDocumentType(unittest.TestCase):
+    def test_default(self) -> None:
+        options: CodecOptions = CodecOptions()
+        obj = options.document_class()
+        obj["a"] = 1
+
+    def test_explicit_document_type(self) -> None:
+        options: CodecOptions[Dict[str, Any]] = CodecOptions()
+        obj = options.document_class()
+        obj["a"] = 1
+
+    def test_typeddict_document_type(self) -> None:
+        options: CodecOptions[Movie] = CodecOptions()
+        # Suppress: Cannot instantiate type "Type[Movie]".
+        obj = options.document_class(name="a", year=1)  # type: ignore[misc]
+        assert obj["year"] == 1
+        assert obj["name"] == "a"
+
+    def test_raw_bson_document_type(self) -> None:
+        options = CodecOptions(RawBSONDocument)
+        doc_bson = b"\x10\x00\x00\x00\x11a\x00\xff\xff\xff\xff\xff\xff\xff\xff\x00"
+        obj = options.document_class(doc_bson)
+        assert len(obj.raw) > 0
+
+    def test_son_document_type(self) -> None:
+        options = CodecOptions(SON[str, Any])
+        obj = options.document_class()
+        obj["a"] = 1
 
 
 if __name__ == "__main__":